fix: Ignore imports with alias
Since aliased imports are not diagnosed (only calls with selector
`maps` and `slices` are checked) we can't safely know if we call any
method in the `exp` package that can't be replaced.

With this change we check both if we ever returned `false` from the
check but even if we didn't (no checks were made) we avoid replacing the
import if it's aliased.